Challenge 3 NaN issue

+++ src/estimator.js
@@ -30,8 +30,8 @@ const covid19ImpactEstimator = (data) => {
   const casesForICUByRequestedTimeSevereImpact = Math.floor(0.05 * Math.floor(infectionsByRequestedTimeSevereImpact));
   const casesForVentilatorsByRequestedTimeImpact = Math.floor(0.02 * Math.floor(infectionsByRequestedTimeImpact));
   const casesForVentilatorsByRequestedTimeSevereIpact = Math.floor(0.02 * Math.floor(infectionsByRequestedTimeSevereImpact));
-  const dollarsInFlightImpact = Math.floor((infectionsByRequestedTimeImpact * parseFloat(incomePopulation)) * parseInt(dailyIncome) * period);
-  const dollarsInFlightSevere = Math.floor((infectionsByRequestedTimeSevereImpact * parseFloat(incomePopulation)) * parseInt(dailyIncome) * period);
+  const dollarsInFlightImpact = Math.floor((infectionsByRequestedTimeImpact * parseFloat(incomePopulation)) * parseFloat(dailyIncome) * period);
+  const dollarsInFlightSevere = Math.floor((infectionsByRequestedTimeSevereImpact * parseFloat(incomePopulation)) * parseFloat(dailyIncome) * period);
 
 
   return {
